Add axe-core accessibility ratchet to the visual-regression job (#1137)
a11y.mjs scans the same six seeded pages as the visual harness, in both
colour schemes (contrast bugs are usually scheme-specific — the #1133
dark-mode banners were exactly this class). Policy: critical/serious
violations are zero-tolerance; moderate/minor are a shrink-only count
ratchet against a11y-baseline.json, mirroring the alert()/inline-script
ratchets. With no baseline committed the scan bootstraps: reports
everything, writes a suggested baseline into the CI artifact, passes.

The fixture seeding is now a shared module (seed.mjs) imported by both
capture.mjs and a11y.mjs; run-a11y.sh boots the server the same way
run-visual.sh does, and the CI job runs the scan right after the visual
compare.

Verified offline against a static page with planted violations: axe
classifies image-alt/label as critical, color-contrast/html-has-lang as
serious, landmark rules as moderate — mapping cleanly onto the
hard-fail / ratchet split.
