+++ Tools/visual-regression/package.json
@@ -5,6 +5,7 @@
   "description": "Visual regression harness: boots chickadee-server, screenshots key pages in light + dark colour schemes, and diffs them against committed baselines.",
   "type": "module",
   "dependencies": {
+    "@axe-core/playwright": "^4.12.1",
     "jszip": "^3.10.1",
     "pixelmatch": "^7.1.0",
     "playwright": "^1.49.1",
